pyproject.toml
[tool.poetry]
name = "flask-api-template"
version = "0.1.0"
description = "This is a Flask API template to create Python backend services"
authors = ["Amirsalar Heydari <amirsalaar@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9.8"
Flask = "^2.0.3"
waitress = "^2.1.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
Flask-Cors = "^3.0.10"
autopep8 = "^1.6.0"
pre-commit = "^2.17.0"
pytest = "^7.1.1"
pytest-mock = "^3.7.0"
pytest-cov = "^3.0.0"
pytest-watch = "^4.2.0"
flake8 = "^4.0.1"
flake8-docstrings = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
